requirements.txt
python-dotenv==0.20.0
python-telegram-bot==13.13
google-api-core==2.10.2
google-cloud-dialogflow==2.16.1
google-cloud-storage==2.6.0
vk-api==11.9.9
